test(opencode): consolidate descriptor suite
